Cmd: [RFK] removed exception handling
I removed Artemis::Exception from Cmd because it introduced issues and
solved none. The most important issue is the fact that DBIC dies
classless and Artemis::Cmd which throws class exceptions. Thus, the
caller had to check for both. Now only classless exceptions are thrown
and the caller can either parse the error message or pass it further up.