[FIX] delete queue with/without jobs
A queue that has jobs attached (no matter if they are finished or not)
should not be deleted. On the other hand empty queues can be
deleted. This distinction is now implemented. Furthermore, a force can
be given to enforce deletion. The caller should make sure that the user
does not abuse this feature.