[FIX] provide command as array to ssh
When cobbler commands and their arguments are provided as array to
Net::OpenSSH->capture (in contrast to one single string) the arguments
are quoted. This means, that a malicious user can no longer cause damage
by providing e.g. a new name like "foo;rm -rf *". This is generally a
good thing.