[FEAT] scenarios in testplans
Scenarios are the way to have testruns that depend on each other
(interdependend testruns), e.g. client-server tests. Yet, Tapper::Cmd
support for scenarios was quite weak. the actual work was still done in
the CLI and not moved into Cmd. This did not allow handling scenarios
other than directly from the CLI.

Now we want to have scenarios in testplans to create testplans with
client-server tests. For this I moved the scenario work into Tapper::Cmd
and called these functions for the testplan handler. This way testplans
now recognise not only multitest but also scenario and thus interdep
testruns.