[FEAT] cancel testrun
This function allows users to cancel a given testrun. It is similar to
the existing function "freehost" only that it expects a testrun id
instead of a host name. If the testrun is running, a message is sent to
the master controller which will end that testrun. If the testrun is not
running, the function does the obvious right thing.