[FIX] add host handled existing hosts correctly
When adding a host that already exists it does not make sense to throw
an exception since the result the user wanted is in fact there.
Also, a user does not see hosts that Tapper keeps as "deleted". If the
user wants to add such a host we need to undelete the host, not throw an
exception.
This patch fixes both.