introduce 'context' to testplans/testruns
The field 'context' of a testplan is passed through
to each testrun's 'program' preconditions. From there
it can be further evaluated.

A context is hash describing what the testruns are
about. That information should be made accessible
to test results (i.e., reports), so that those
can be easily found and sorted back together.